Fixes "Skip" buttons in .mobi
Adds article-header div as a wrapper on title, domain, and other fields on top of the article.

Renames link to cssLink and a to skip.

saveArticle function takes additional - rather self-explanatory - bool addSkipButton argument. It can be used to disable "Skip" button on user's demand or to disable it in the last article (which has nowhere to point to).

Comments out CSS code describing the first letter in the article.